weather-ai-app/requirements.txt

# Core
Flask==3.0.3
python-dotenv==1.0.1
gunicorn==22.0.0

# Data Science / ML
tensorflow==2.19.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
joblib==1.4.2

# Visualization
plotly==5.22.0
